contact_finder/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
gspread>=5.12.0
google-auth>=2.23.0
googlesearch-python>=1.2.0
